@@ -1,7 +1,7 @@
 {
   "name" : "hamhamfonfon/astrobin-ws",
   "license": "MIT",
-  "version": "1.0.1",
+  "version": "2.0.0",
   "description" : "Astrobin WebServices for PHP projects",
   "type" : "library",
   "authors" : [{
@@ -15,7 +15,10 @@
     "issues": "https://github.com/HamHamFonFon/Astrobin-API-PHP/issues"
   },
   "require" : {
-    "php": ">7.0.0"
+    "php": ">=7.3.0",
+    "guzzlehttp/guzzle": "^7.0",
+    "ext-ctype": "*",
+    "ext-json": "*"
   },
   "autoload" : {
     "psr-4" : {
@@ -27,7 +30,8 @@
   },
   "target-dir" : "",
   "require-dev": {
-    "phpunit/phpunit": "^6",
-    "squizlabs/php_codesniffer": "^3.2"
+    "squizlabs/php_codesniffer": "^3.2",
+    "phpunit/phpunit": "^9.3",
+    "dg/bypass-finals": "^1.3"
   }
 }
