@@ -1,29 +1,46 @@
 <?php
 
-namespace Astrobin;
+declare(strict_types=1);
 
-use Astrobin\Exceptions\WsException;
+namespace AstrobinWs;
+
+use AstrobinWs\Exceptions\WsException;
+use AstrobinWs\Exceptions\WsResponseException;
+use AstrobinWs\Response\AstrobinResponse;
+use AstrobinWs\Response\Collection;
+use AstrobinWs\Response\Image;
+use AstrobinWs\Response\ListCollection;
+use AstrobinWs\Response\ListImages;
+use AstrobinWs\Response\Today;
+use GuzzleHttp\Client;
+use GuzzleHttp\Exception\GuzzleException;
+use http\Client\Response;
+use Psr\Http\Message\ResponseInterface;
+use Psr\Http\Message\StreamInterface;
 
 /**
  * Class AstrobinWebService
+ *
  * @package AppBundle\Astrobin
  */
 abstract class AbstractWebService
 {
-    const ASTROBIN_URL = 'https://www.astrobin.com/api/v1/';
-    const MAX_REDIRS = 10;
-    const LIMIT_MAX = 20;
-    const TIMEYOUT = 30;
-
-    const METHOD_GET = 'GET';
-    const METHOD_POST = 'POST';
-    const METHOD_PUT = 'PUT';
+    public const LIMIT_MAX = 20;
+    public const TIMEOUT = 30;
 
     protected $timeout;
     private $apiKey;
     private $apiSecret;
-    /** @var CurlHttpRequestInterface */
-    protected $curlRequest;
+
+    protected static $headers = [
+        'Accept' => GuzzleSingleton::APPLICATION_JSON,
+        'Content-Type' => GuzzleSingleton::APPLICATION_JSON
+    ];
+
+    /**
+     * @var Client
+     */
+    private $client;
 
     /**
      * AbstractWebService constructor.
@@ -32,193 +49,227 @@ abstract class AbstractWebService
     {
         $this->apiKey = getenv('ASTROBIN_API_KEY');
         $this->apiSecret = getenv('ASTROBIN_API_SECRET');
-        $this->timeout = self::TIMEYOUT;
+        $this->timeout = self::TIMEOUT;
+        $this->buildFactory();
     }
 
 
     /**
-     * @param $endPoint
-     * @param $method
-     * @param $data
-     * @return mixed|null
-     * @throws WsException
+     * Get Guzzle Instance
      */
-    protected function call($endPoint, $method, $data): object
+    public function buildFactory(): void
     {
-        if (is_null($this->apiKey) || is_null($this->apiSecret)) {
-            throw new WsException(sprintf("Astrobin Webservice : API key or API secret are null"));
-        }
-
-        $urlAstrobin = $this->buildUrl($endPoint, $data);
-        /** @var CurlHttpRequestInterface curlRequest */
-        $this->curlRequest = new CurlHttpRequest();
-        $options = $this->initCurlOptions($method, $urlAstrobin);
-        $this->curlRequest->setOptionArray($options);
-
-        if (!$resp = $this->curlRequest->execute()) {
-            if (empty($resp)) {
-                $dataErr = (!is_array($data)) ? [$data] : $data;
-                throw new WsException(sprintf("[Astrobin Response] Empty response from \"%s\", check data : %s", $endPoint, implode(' . ', $dataErr)));
-            }
-            // show problem and throw exception
-            throw new WsException(
-                sprintf("[Astrobin Response] HTTP Error (curl_exec) #%u: %s", $this->curlRequest->getErrNo(), $this->curlRequest->getError())
-            );
-        }
-        $this->curlRequest->close();
-
-        if (!$resp || empty($resp)) {
-            throw new WsException("[Astrobin Response] Empty Json");
-        }
-
-        return $this->buildResponse($resp);
+        $this->client = GuzzleSingleton::getInstance();
     }
 
+    /**
+     * Return endpoint
+     */
+    abstract protected function getEndPoint(): string;
 
     /**
-     * DEBUG request
-     * TODO : make better, this is only for a specific dev
+     * Get Instance of response object Entity
      *
-     * @param $endPoint
-     * @param $method
-     * @param $data
-     * @return \stdClass
-     */
-    protected function callDebug($endPoint, $method, $data)
-    {
-        $returnErr = new \stdClass();
-
-        $urlAstrobin = $this->buildUrl($endPoint, $data);
-        $this->curlRequest = new CurlHttpRequest();
-        $options = $this->initCurlOptions($method, $urlAstrobin);
-        $this->curlRequest->setOptionArray($options);
-
-        if (!$resp = $this->curlRequest->execute()) {
-
-            $returnErr->http_code = curl_getinfo($this->curlRequest->getHandle(), CURLINFO_HTTP_CODE);
-            $returnErr->endpoint = $endPoint;
-            $returnErr->method = $endPoint;
-            $returnErr->data = $data;
-        }
-
-        $this->curlRequest->close();
-
-        return $returnErr;
-    }
-
-    /**
-     * Build the WebService URL
-     * @param $endPoint
-     * @param $data
      * @return string
      */
-    private function buildUrl($endPoint, $data): string
+    abstract protected function getObjectEntity(): ?string;
+
+    /**
+     * Get instance of response collection entity
+     * @return string
+     */
+    abstract protected function getCollectionEntity(): ?string;
+
+    /**
+     * Build EndPoint
+     *
+     * @param string|null $param
+     *
+     * @return string
+     */
+    private function buildEndpoint(?string $param): string
     {
-        // Build URL with params
-        $url = self::ASTROBIN_URL . $endPoint;
+        return (!is_null($param)) ? sprintf('/api/v1/%s/%s', $this->getEndPoint(), $param) : sprintf('/api/v1/%s', $this->getEndPoint());
+    }
 
-        if (is_array($data) && 0 < count($data)) {
-            $paramData = implode('&', array_map(function ($k, $v) {
-                $formatValue = "%s";
-                if (is_numeric($v)) {
-                    $formatValue = "%d";
-                }
-                return sprintf("%s=$formatValue", $k, $v);
-            }, array_keys($data), $data));
+    /**
+     * @param string|null $id
+     * @param array|null $queryParams
+     *
+     * @return \stdClass|null
+     * @throws WsException
+     * @throws \JsonException
+     */
+    protected function get(?string $id, ?array $queryParams): ?string
+    {
+        return $this->buildRequest($id, null, $queryParams, null, GuzzleSingleton::METHOD_GET);
+    }
 
-            $url .= '?' . $paramData;
-        } else {
-            if ('/' !== substr($url, strlen($url)-1, strlen($url))) {
-                $url .= '/';
-            }
-            // Warning : the "/" before "?" is mandatory or else no response from WS...
-            $url .= $data . '/?';
+    /**
+     * NOT USED, just for example
+     *
+     * @param string $id
+     * @param array|null $queryParams
+     * @param array|null $body
+     *
+     * @return \stdClass|null
+     * @throws WsException
+     * @throws \JsonException
+     */
+    protected function post(string $id, ?array $queryParams, ?array $body): ?string
+    {
+        $this->buildRequest($id, $body, $queryParams, null, GuzzleSingleton::METHOD_POST);
+    }
+
+    /**
+     * @param string|null $id
+     * @param array|null $body
+     * @param array|null $queryParams
+     * @param array|null $headers
+     * @param string $method
+     *
+     * @return \stdClass|null
+     * @throws WsException
+     * @throws \JsonException
+     */
+    private function buildRequest(?string $id, ?array $body, ?array $queryParams, ?array $headers, string $method): ?string
+    {
+        if (is_null($this->apiKey) || is_null($this->apiSecret)) {
+            throw new WsException(WsException::KEYS_ERROR, 500, null);
         }
 
-        // Add keys and format
-        $params = [
+        $endPoint = $this->buildEndpoint($id);
+        if (!is_null($headers)) {
+            $options['headers'] = array_merge(self::$headers, $headers);
+        } else {
+            $options['headers'] = self::$headers;
+        }
+
+        $astrobinParams = [
             'api_key' => $this->apiKey,
             'api_secret' => $this->apiSecret,
             'format' => 'json'
         ];
 
-        $httpParams = implode('', array_map(function ($k, $v) {
-            return sprintf("&%s=%s", $k, $v);
-        }, array_keys($params), $params));
-        $url .= $httpParams;
-
-        return $url;
-    }
-
-
-    /**
-     * Options for cURL request
-     *
-     * @param $method
-     * @param string $url
-     * @return mixed
-     */
-    protected function initCurlOptions($method, $url): array
-    {
-        // Options CURL
-        $options = [
-            CURLOPT_URL => $url,
-            CURLOPT_RETURNTRANSFER => true,
-            CURLOPT_ENCODING => "",
-            CURLOPT_MAXREDIRS => self::MAX_REDIRS,
-            CURLOPT_HEADER => "Accept:application/json",
-            CURLOPT_CONNECTTIMEOUT => $this->timeout,
-            CURLOPT_HTTP_VERSION => CURL_HTTP_VERSION_1_1,
-            CURLOPT_TIMEOUT => $this->timeout,
-            CURLOPT_SSL_VERIFYPEER => false
-        ];
-
-        // GET
-        if (self::METHOD_GET === $method) {
-            $options = array_replace_recursive($options, [
-                CURLOPT_CUSTOMREQUEST => self::METHOD_GET,
-                CURLOPT_HTTPGET => true,
-            ]);
+        if (is_null($queryParams)) {
+            $queryParams = [];
         }
 
-        return $options;
+        $options = [
+            'query' => array_filter(array_merge($astrobinParams, $queryParams))
+        ];
+
+        if (!is_null($body) && !empty($body)) {
+            $options['body'] = $body;
+        }
+
+        try {
+            /**
+             * @var ResponseInterface $responseGuzzle
+             */
+            $responseGuzzle = $this->client->request($method, $endPoint, $options);
+        } catch (GuzzleException $e) {
+            //throw
+            $responseGuzzle = null;
+//            throw new WsException($e->getMessage(), $e->getCode(), $e);
+        }
+
+        return $this->getResponse($responseGuzzle);
     }
 
 
     /**
      * Check response and jsondecode object
      *
-     * @param $resp
-     * @return object|null
+     * @param ResponseInterface $response
+     *
+     * @return mixed|null
      * @throws WsException
      */
-    public function buildResponse($resp): object
+    public function getResponse(ResponseInterface $response): string
     {
-        $obj = null;
-
-        if (is_string($resp)) {
-            if (false === strpos($resp, '{', 0)) {
-                // check if html
-                if (false !== strpos($resp, '<html', 0)) {
-                    throw new WsException(sprintf("[Astrobin Response] Response in HTML format :\n %s", $resp));
-                }
-                throw new WsException(sprintf("[Astrobin Response] Not a JSON valid format :\n %s", $resp));
-            }
-            $obj = json_decode($resp);
-            if (JSON_ERROR_NONE != json_last_error()) {
-                throw new WsException(
-                    sprintf("[Astrobin ERROR] Error JSON :\n%s", json_last_error())
-                );
-            }
-            if (array_key_exists('error', $obj)) {
-                throw new WsException(
-                    sprintf("[Astrobin ERROR] Response : %s", $obj->error)
-                );
-            }
-        } else {
-            throw new WsException("[Astrobin ERROR] Response is not a string, got ". gettype($resp) . " instead.");
+        if (200 !== $response->getStatusCode()) {
+            throw new WsException(sprintf(WsException::GUZZLE_RESPONSE, $response->getReasonPhrase()), 500, null);
         }
 
-        return $obj;
+        /**
+         * @var StreamInterface $body
+        */
+        $body = $response->getBody();
+
+        if (false === $body->isReadable()) {
+            throw new WsException(WsException::ERR_READABLE, 500, null);
+        }
+
+        if (0 === $body->getSize()) {
+            throw new WsException(sprintf(WsException::ERR_EMPTY, $this->getEndPoint()), 500, null);
+        }
+
+        $contents = $body->getContents();
+        if (false === strpos($contents, '{', 0)) {
+            throw new WsException(sprintf(WsException::ERR_JSON, (string)$body), 500, null);
+        }
+
+        return $contents;
+    }
+
+    /**
+     * Convert string into Json
+     *
+     * @param string $contents
+     *
+     * @return \stdClass
+     * @throws \JsonException
+     * @throws WsResponseException
+     */
+    protected function deserialize(string $contents): \stdClass
+    {
+        $responseJson = json_decode($contents, false, 512, JSON_THROW_ON_ERROR);
+        if (property_exists($responseJson, "objects") && property_exists($responseJson, "meta") && 0 === $responseJson->meta->total_count) {
+            throw new WsResponseException(WsException::RESP_EMPTY, 500, null);
+        }
+
+        return $responseJson;
+    }
+
+    /**
+     * Build response from WebService Astrobin
+     *
+     * @param string $response
+     *
+     * @return AstrobinResponse
+     * @throws WsResponseException
+     * @throws \ReflectionException|\JsonException
+     */
+    protected function buildResponse(string $response): ?AstrobinResponse
+    {
+        $astrobinResponse = null;
+        $object = $this->deserialize($response);
+
+        /** @var Image|Today|Collection|AstrobinResponse $entity */
+        $entity = $this->getObjectEntity();
+
+        /** @var ListImages|ListCollection|ListImages|AstrobinResponse $collectionEntity */
+        $collectionEntity = $this->getCollectionEntity();
+
+        if (property_exists($object, "objects") && 0 < count($object->objects)) {
+            $listObjects = $object->objects;
+            if (1 < count($listObjects)) {
+                $astrobinResponse = new $collectionEntity();
+                foreach ($listObjects as $object) {
+                    $entity = new $entity();
+                    $entity->fromObj($object);
+                    $astrobinResponse->add($entity);
+                }
+            } else {
+                $astrobinResponse = new $entity();
+                $astrobinResponse->fromObj(reset($listObjects));
+            }
+        } else {
+            $astrobinResponse = new $entity();
+            $astrobinResponse->fromObj($object);
+        }
+
+        return $astrobinResponse;
     }
 }
