@@ -1,12 +1,14 @@
 <?php
 
-namespace Astrobin\Exceptions;
+declare(strict_types=1);
+
+namespace AstrobinWs\Exceptions;
 
 /**
  * Class WsResponseException
+ *
  * @package Astrobin\Exceptions
  */
-class WsResponseException extends WsException
+final class WsResponseException extends WsException
 {
-
 }
