@@ -1,8 +1,10 @@
 <?php
 
-namespace Astrobin\Response;
+declare(strict_types=1);
 
-use Astrobin\Exceptions\WsResponseException;
+namespace AstrobinWs\Response;
+
+use AstrobinWs\Exceptions\WsResponseException;
 
 /**
  * Class AbstractResponse
@@ -17,7 +19,7 @@ abstract class AbstractResponse
      * @throws WsResponseException
      * @throws \ReflectionException
      */
-    public function fromObj(\stdClass $obj)
+    public function fromObj(\stdClass $obj): void
     {
         $this->fromArray((array)$obj);
     }
@@ -27,12 +29,14 @@ abstract class AbstractResponse
      * Build properties of class based on WS response
      * @param array $objArr
      * @throws WsResponseException
-     * @throws \ReflectionException
      */
-    private function fromArray(array $objArr)
+    private function fromArray(array $objArr): void
     {
         $listNotFields = ['listImages'];
-        /** @var \ReflectionClass $reflector */
+
+        /**
+         * @var \ReflectionClass $reflector
+         */
         $reflector = new \ReflectionClass($this);
         foreach ($reflector->getProperties() as $property) {
             if (in_array($property->getName(), $listNotFields)) {
@@ -42,7 +46,9 @@ abstract class AbstractResponse
             }
             if (!array_key_exists($property->getName(), $objArr)) {
                 throw new WsResponseException(
-                    sprintf("Property \"%s\" doesn't exist in class %s", $property->getName(), get_class($this))
+                    sprintf("Property \"%s\" doesn't exist in class %s", $property->getName(), get_class($this)),
+                    500,
+                    null
                 );
             }
 
