@@ -1,21 +1,23 @@
 <?php
 
-namespace Astrobin\Response\Iterators;
+declare(strict_types=1);
+
+namespace AstrobinWs\Response\Iterators;
 
 /**
  * Class ImageIterator
  * @package Astrobin\Response
  */
-class ImageIterator implements \Iterator
+final class ImageIterator implements \Iterator
 {
-
+    /** @var array  */
     private $var = [];
 
     /**
      * ImageIterator constructor.
      * @param $array
      */
-    public function __construct($array)
+    public function __construct(array $array)
     {
         if (is_array($array)) {
             $this->var = $array;
@@ -51,11 +53,10 @@ class ImageIterator implements \Iterator
     /**
      * @return bool
      */
-    public function valid()
+    public function valid(): bool
     {
         $key = key($this->var);
-        $var = (!is_null($key) && false !== $key);
-        return $var;
+        return !is_null($key) && false !== $key;
     }
 
 
