@@ -1,13 +1,16 @@
 <?php
 
-namespace Astrobin\Services;
+declare(strict_types=1);
 
-use Astrobin\AbstractWebService;
-use Astrobin\Exceptions\WsException;
-use Astrobin\Exceptions\WsResponseException;
-use Astrobin\Response\Collection;
-use Astrobin\Response\Image;
-use Astrobin\Response\ListCollection;
+namespace AstrobinWs\Services;
+
+use AstrobinWs\Response\AstrobinResponse;
+use AstrobinWs\AbstractWebService;
+use AstrobinWs\Exceptions\WsException;
+use AstrobinWs\Exceptions\WsResponseException;
+use AstrobinWs\Response\Collection;
+use AstrobinWs\Response\Image;
+use AstrobinWs\Response\ListCollection;
 
 /**
  * Class GetCollection
@@ -15,131 +18,93 @@ use Astrobin\Response\ListCollection;
  */
 class GetCollection extends AbstractWebService implements WsInterface
 {
+    use WsAstrobinTrait;
 
-    const END_POINT = 'collection/';
-
+    private const END_POINT = 'collection';
 
     /**
-     * @param null $id
-     * @return Collection|null
-     * @throws WsException
-     * @throws WsResponseException
-     * @throws \ReflectionException
+     * @return string
      */
-    public function getCollectionById($id = null): Collection
+    protected function getEndPoint(): string
     {
-        if (is_null($id) || empty($id)) {
-            throw new WsException('Astrobon Webservice Collection : id empty');
-        }
+        return self::END_POINT;
+    }
 
-        $astrobinCollection = $this->callWs($id);
-        if (isset($astrobinCollection->images) && 0 < count($astrobinCollection->images)) {
-            $astrobinCollection = $this->getImagesCollection($astrobinCollection);
-        }
+    /**
+     * @return string
+     */
+    protected function getObjectEntity(): string
+    {
+        return Collection::class;
+    }
 
-        return $astrobinCollection;
+    /**
+     * @return string
+     */
+    protected function getCollectionEntity(): string
+    {
+        return ListCollection::class;
     }
 
 
     /**
-     * @param null $username
-     * @param null $limit
-     * @return ListCollection|null
+     * Only for retro-compatibility with version 1.x
+     *
+     * @param $id
+     *
+     * @return AstrobinResponse
+     * @throws WsException
+     * @throws \ReflectionException|\JsonException
+     */
+    public function getCollectionById($id): ?AstrobinResponse
+    {
+        return $this->getById($id);
+    }
+
+    /**
+     * @param string|null $id
+     *
+     * @return AstrobinResponse
      * @throws WsException
      * @throws WsResponseException
+     * @throws \JsonException
      * @throws \ReflectionException
      */
-    public function getListCollectionByUser($username = null, $limit = null): ListCollection
+    public function getById(?string $id): ?AstrobinResponse
+    {
+        if (is_null($id)) {
+            throw new WsResponseException(sprintf(WsException::EMPTY_ID, $id), 500, null);
+        }
+
+        $response = $this->get($id, null);
+        $collection = $this->buildResponse($response);
+
+        if (!is_null($collection)) {
+            return $this->getImagesFromResource($collection);
+        }
+
+        return null;
+    }
+
+
+    /**
+     * @param string|null $username
+     * @param int|null $limit
+     *
+     * @return ListCollection|null
+     * @throws WsException
+     * @throws \ReflectionException
+     * @throws \JsonException
+     */
+    public function getListCollectionByUser(?string $username, ?int $limit): ?AstrobinResponse
     {
         if (parent::LIMIT_MAX < $limit) {
             $limit = parent::LIMIT_MAX;
         }
         $params = ['user' => $username, 'limit' => $limit];
         /** @var ListCollection $astrobinListCollection */
-        $astrobinListCollection = $this->callWs($params);
+        $response = $this->get(null, $params);
 
-        return $astrobinListCollection;
-    }
-
-
-    /**
-     * Retrieve images of a collection by WS GetImage
-     * @param Collection $astrobinCollection
-     * @return Collection
-     * @throws WsException
-     * @throws WsResponseException
-     * @throws \ReflectionException
-     */
-    private function getImagesCollection(Collection $astrobinCollection): Collection
-    {
-        $listImagesId = array_map(function ($path) {
-            if (preg_match('/\/([\d]+)/', $path, $matches)) {
-                return $matches[1];
-            }
-        }, $astrobinCollection->images);
-        if (0 < count($listImagesId)) {
-            foreach ($listImagesId as $imageId) {
-                $imgRawCall = $this->call(GetImage::END_POINT, parent::METHOD_GET, $imageId);
-
-                $image = new Image();
-                $image->fromObj($imgRawCall);
-
-                $astrobinCollection->add($image);
-            }
-        }
-
-        return $astrobinCollection;
-    }
-
-    /**
-     * @param array|integer $params
-     * @return ListCollection|Collection|null
-     * @throws WsException
-     * @throws WsResponseException
-     * @throws \ReflectionException
-     */
-    public function callWs($params = [])
-    {
-        $rawResp = $this->call(self::END_POINT, parent::METHOD_GET, $params);
-        if (!is_object($rawResp)) {
-            throw new WsResponseException("Response from Astrobin is empty");
-        } else {
-            if (property_exists($rawResp, "objects") && property_exists($rawResp, "meta") && 0 < $rawResp->meta->total_count) {
-                return $this->responseWs($rawResp->objects);
-            } else {
-                return $this->responseWs([$rawResp]);
-            }
-        }
-    }
-
-
-    /**
-     * @param $objects
-     * @return Collection|null
-     * @throws WsException
-     * @throws WsResponseException
-     * @throws \ReflectionException
-     */
-    public function responseWs($objects = [])
-    {
-        $astrobinResponse = null;
-        if (is_array($objects) && 0 < count($objects)) {
-            if (1 < count($objects)) {
-                /** @var ListCollection $astrobinResponse */
-                $astrobinResponse = new ListCollection();
-                foreach ($objects as $object) {
-                    $collection = new Collection();
-                    $collection->fromObj($object);
-
-                    $this->getImagesCollection($collection);
-                    $astrobinResponse->add($collection);
-                }
-            } else {
-                /** @var Collection $astrobinResponse */
-                $astrobinResponse = new Collection();
-                $astrobinResponse->fromObj($objects[0]);
-            }
-        }
-        return $astrobinResponse;
+        return $this->buildResponse($response);
     }
 }
