@@ -1,46 +1,84 @@
 <?php
 
-namespace Astrobin\Services;
+declare(strict_types=1);
 
-use Astrobin\AbstractWebService;
-use Astrobin\Exceptions\WsException;
-use Astrobin\Exceptions\WsResponseException;
-use Astrobin\Response\Collection;
-use Astrobin\Response\Image;
-use Astrobin\Response\ListImages;
+namespace AstrobinWs\Services;
+
+use AstrobinWs\AbstractWebService;
+use AstrobinWs\Exceptions\WsException;
+use AstrobinWs\Exceptions\WsResponseException;
+use AstrobinWs\Response\AstrobinResponse;
+use AstrobinWs\Response\Collection;
+use AstrobinWs\Response\Image;
+use AstrobinWs\Response\ListImages;
 
 /**
  * Class GetImage
+ *
  * @package Astrobin\Services
  */
 class GetImage extends AbstractWebService implements WsInterface
 {
-
-    const END_POINT = 'image/';
-
+    public const END_POINT = 'image';
 
     /**
-     * @param $id
-     * @return Image|null
-     * @throws WsResponseException
-     * @throws \Astrobin\Exceptions\WsException
-     * @throws \ReflectionException
+     * @return string
      */
-    public function getImageById($id): Image
+    protected function getEndPoint(): string
     {
-        if (is_null($id) || !ctype_alnum($id)) {
-            throw new WsResponseException(sprintf("[Astrobin response] '%s' is not a correct value, alphanumeric expected", $id));
-        }
-        return $this->callWs($id);
+        return self::END_POINT;
     }
 
-    public function debugImageById($id)
+    /**
+     * @return string
+     */
+    protected function getObjectEntity(): string
     {
-        if (is_null($id) || !ctype_alnum($id)) {
-            throw new WsResponseException(sprintf("[Astrobin response] '%s' is not a correct value, alphanumeric expected", $id));
-        }
+        return Image::class;
+    }
 
-        return $this->callDebug(self::END_POINT, AbstractWebService::METHOD_GET, $id);
+    /**
+     * @return string
+     */
+    protected function getCollectionEntity(): string
+    {
+        return ListImages::class;
+    }
+
+    /**
+     * Only for retro-compatibility with version 1.x
+     *
+     * @param $id
+     *
+     * @return AstrobinResponse|null
+     * @throws WsException
+     * @throws WsResponseException
+     * @throws \ReflectionException
+     * @throws \JsonException
+     */
+    public function getImageById(string $id): ?AstrobinResponse
+    {
+        return $this->getById($id);
+    }
+
+    /**
+     * Id of image can be a string or an int
+     *
+     * @param string|null $id
+     *
+     * @return AstrobinResponse
+     * @throws WsException
+     * @throws WsResponseException
+     * @throws \ReflectionException
+     * @throws \JsonException
+     */
+    public function getById(?string $id): ?AstrobinResponse
+    {
+        if (is_null($id)) {
+            throw new WsResponseException(sprintf(WsException::EMPTY_ID, $id), 500, null);
+        }
+        $response = $this->get($id, null);
+        return $this->buildResponse($response);
     }
 
     /**
@@ -48,91 +86,110 @@ class GetImage extends AbstractWebService implements WsInterface
      *
      * @param $subjectId
      * @param $limit
+     *
      * @return ListImages|Image|null
      * @throws WsResponseException
-     * @throws \Astrobin\Exceptions\WsException
+     * @throws WsException
      * @throws \ReflectionException
+     * @throws \JsonException
      */
-    public function getImagesBySubject($subjectId, $limit)
+    public function getImagesBySubject(string $subjectId, int $limit): ?AstrobinResponse
     {
         if (parent::LIMIT_MAX < $limit) {
             return null;
         }
 
         $params = ['subjects' => $subjectId, 'limit' => $limit];
-        return $this->callWs($params);
+        $response = $this->get(null, $params);
+        return $this->buildResponse($response);
     }
 
 
     /**
      * @param $description
      * @param $limit
+     *
      * @return ListImages|Image|null
      * @throws WsResponseException
-     * @throws \Astrobin\Exceptions\WsException
+     * @throws WsException
      * @throws \ReflectionException
+     * @throws \JsonException
      */
-    public function getImagesByDescription($description, $limit)
+    public function getImagesByDescription(string $description, int $limit): ?AstrobinResponse
     {
         if (parent::LIMIT_MAX < $limit) {
             return null;
         }
 
         $params = ['title__icontains' => urlencode($description), 'limit' => $limit];
-        return $this->callWs($params);
+        $response = $this->get(null, $params);
+        return $this->buildResponse($response);
     }
 
 
     /**
      * Return an Collection per user name
+     *
      * @param $userName
      * @param $limit
+     *
      * @return ListImages|Image|null
      * @throws WsResponseException
-     * @throws \Astrobin\Exceptions\WsException
+     * @throws WsException
      * @throws \ReflectionException
+     * @throws \JsonException
      */
-    public function getImagesByUser($userName, $limit)
+    public function getImagesByUser(string $userName, int $limit): ?AstrobinResponse
     {
         if (parent::LIMIT_MAX < $limit) {
             return null;
         }
 
         $params = ['user' => $userName, 'limit' => $limit];
-        return $this->callWs($params);
+        $response = $this->get(null, $params);
+
+        return $this->buildResponse($response);
     }
 
 
     /**
-     * @param null $dateFromStr
-     * @param null $dateToStr
+     * @param string|null $dateFromStr
+     * @param string|null $dateToStr
+     *
      * @return ListImages|Image|null
+     * @throws WsException
      * @throws WsResponseException
-     * @throws \Astrobin\Exceptions\WsException
      * @throws \ReflectionException
+     * @throws \Exception
      */
-    public function getImagesByRangeDate($dateFromStr = null, $dateToStr = null)
+    public function getImagesByRangeDate(?string $dateFromStr, ?string $dateToStr): ?AstrobinResponse
     {
-
         if (is_null($dateToStr)) {
-            /** @var \DateTime $dateTo */
+            /**
+             * @var \DateTimeInterface $dateTo
+            */
             $dateTo = new \DateTime('now');
         } else {
-            /** @var \DateTime $dateTo */
+            /**
+             * @var \DateTimeInterface $dateTo
+            */
             $dateTo = new \DateTime($dateToStr);
         }
 
         if (false === strtotime($dateFromStr)) {
-            throw new WsException(sprintf("Format \"%s\" is not a correct format, please use YYYY-mm-dd", $dateFromStr));
+            throw new WsException(sprintf(WsException::ERR_DATE_FORMAT, $dateFromStr), 500, null);
         }
 
-        /** @var \DateTime $dateFrom */
+        /**
+         * @var \DateTime $dateFrom
+        */
         $dateFrom = \DateTime::createFromFormat('Y-m-d', $dateFromStr);
-        if (false === $dateFrom->format('Y-m-d') || $dateFromStr !== $dateFrom->format('Y-m-d')) {
-            throw new WsException(sprintf("Format \"%s\" is not a correct format for a date, please use YYYY-mm-dd instead", $dateFromStr));
+        if ($dateFromStr !== $dateFrom->format('Y-m-d')) {
+            throw new WsException(sprintf(WsException::ERR_DATE_FORMAT, $dateFromStr), 500, null);
+        }
 
-        } elseif (false !== $dateFrom && array_sum($dateFrom->getLastErrors())) {
-            throw new WsException("Error date format : \n" . print_r($dateFrom->getLastErrors()));
+        if (false !== $dateFrom && array_sum($dateFrom->getLastErrors())) {
+            throw new WsException(WsException::ERR_DATE . print_r($dateFrom->getLastErrors()), 500, null);
         }
 
         $params = [
@@ -141,65 +198,7 @@ class GetImage extends AbstractWebService implements WsInterface
             'limit' => AbstractWebService::LIMIT_MAX
         ];
 
-        return $this->callWs($params);
-    }
-
-    /**
-     * Call WS "image" with parameters
-     * @param array $params
-     * @return ListImages|Image|null
-     * @throws WsResponseException
-     * @throws \Astrobin\Exceptions\WsException
-     * @throws \ReflectionException
-     */
-    public function callWs($params = [])
-    {
-        $rawResp = $this->call(self::END_POINT, AbstractWebService::METHOD_GET, $params);
-        if (!is_object($rawResp)) {
-            throw new WsResponseException("Response from Astrobin is empty");
-        } else {
-            if (property_exists($rawResp, "objects") && property_exists($rawResp, "meta")) {
-                if (0 < $rawResp->meta->total_count) {
-                    return $this->responseWs($rawResp->objects);
-                }
-                throw new WsResponseException(sprintf("Astrobin doen't find any objects with params : %s", json_encode($params)));
-            } else {
-                return $this->responseWs([$rawResp]);
-            }
-        }
-        return null;
-    }
-
-
-    /**
-     * Build response from WebService Astrobin
-     * @param array $objects
-     * @return ListImages|Image|null
-     * @throws WsResponseException
-     * @throws \ReflectionException
-     */
-    public function responseWs($objects = [])
-    {
-        $astrobinResponse = null;
-        if (is_array($objects) && 0 < count($objects)) {
-            if (1 < count($objects)) {
-                /** @var Collection $astrobinCollection */
-                $astrobinResponse = new ListImages();
-                foreach ($objects as $object) {
-                    $image = new Image();
-                    $image->fromObj($object);
-                    $astrobinResponse->add($image);
-                }
-            } else {
-                /** @var Image $astrobinResponse */
-                $astrobinResponse = new Image();
-                $astrobinResponse->fromObj($objects[0]);
-            }
-        } else {
-            $astrobinResponse = new Image();
-            $astrobinResponse->fromObj($objects[0]);
-        }
-
-        return $astrobinResponse;
+        $response = $this->get(null, $params);
+        return $this->buildResponse($response);
     }
 }
